% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CancerSimulationR.R
\docType{package}
\name{CancerSimulationR}
\alias{CancerSimulationR}
\alias{CancerSimulationR-package}
\title{CancerSimulationR - Interactive cancer simulation activity in R}
\description{
Shiny application ???
}
