% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch-GUI.R
\name{LaunchGUI}
\alias{LaunchGUI}
\title{GUI for launching shiny Simulation Application}
\usage{
LaunchGUI()
}
\description{
GUI for launching shiny Simulation Application
}
\examples{
\dontrun{LaunchGUI()}
}
